+++ app/Concerns/Publishable.php
@@ -1,0 +1,36 @@
+<?php
+
+namespace App\Concerns;
+
+use Illuminate\Support\Carbon;
+
+trait Publishable
+{
+    public function initializePublishable()
+    {
+        $this->casts['is_published'] = 'boolean';
+        $this->casts['published_at'] = 'datetime';
+    }
+
+    public function getIsPublishedAttribute()
+    {
+        return (bool) optional($this->published_at)->isPast();
+    }
+
+    public function setIsPublishedAttribute(bool $isPublished)
+    {
+        if ($this->is_published === $isPublished) {
+            return;
+        }
+        if (!$isPublished) {
+            return $this->attributes['published_at'] = null;
+        }
+
+        $this->attributes['published_at'] = Carbon::now();
+    }
+
+    public function scopePublished($query)
+    {
+        return $query->where('published_at', '<=', Carbon::now());
+    }
+}